app/requirements.txt
opentelemetry-instrumentation-requests==0.45b0
opentelemetry-instrumentation-flask==0.45b0
opentelemetry-distro==0.45b0
opentelemetry-exporter-otlp==1.24.0
opentelemetry-sdk==1.24.0
opentelemetry-api==1.24.0
flask==3.0.3